+++ lib/services/lastfm.ts
@@ -1,4 +1,4 @@
-export * from "./lastfm/getRecentTracks";
-export * from "./lastfm/getTopAlbums";
-export * from "./lastfm/getTopArtists";
-export * from "./lastfm/getTopTracks";
+export * from "./lastfm/user/get-recent-tracks";
+export * from "./lastfm/user/get-top-albums";
+export * from "./lastfm/user/get-top-artists";
+export * from "./lastfm/user/get-top-tracks";
